+++ crates/pine-builtins/src/namespaces/types.rs
@@ -14,6 +14,7 @@ pub(crate) const PLOT: PineType = PineType::new(Qualifier::Const, ValueKind::Plo
 pub(crate) const HLINE: PineType = PineType::new(Qualifier::Const, ValueKind::HLine);
 pub(crate) const SERIES_LABEL: PineType = PineType::new(Qualifier::Series, ValueKind::Label);
 pub(crate) const SERIES_LINE: PineType = PineType::new(Qualifier::Series, ValueKind::Line);
+pub(crate) const SERIES_LINE_FILL: PineType = PineType::new(Qualifier::Series, ValueKind::LineFill);
 pub(crate) const SERIES_BOX: PineType = PineType::new(Qualifier::Series, ValueKind::Box);
 pub(crate) const SERIES_TABLE: PineType = PineType::new(Qualifier::Series, ValueKind::Table);
 pub(crate) const VOID: PineType = PineType::new(Qualifier::Const, ValueKind::Void);